fix: returns project user. Not User list

+++ back/src/main/java/com/experis/lagalt/services/ProjectService.java
@@ -6,7 +6,6 @@ import com.experis.lagalt.repositories.ProjectRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -34,15 +33,17 @@ public class ProjectService {
     }
 
     public void deleteProject(long id) {
-        projectRepository.deleteById(id);
+        if(projectExists(id)) {
+            projectRepository.deleteById(id);
+        }
     }
 
-    public List<User> getProjectUsers(long id) {
-        ArrayList<User> users = new ArrayList<>();
+    public User getProjectUser(long id) {
+        User projectUser = new User();
         if (projectExists(id)) {
             Project project = findProject(id);
-            users.addAll(project.getUsers());
+            projectUser = project.getUser();
         }
-        return users;
+        return projectUser;
     }
 }
